backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
redis==5.0.8
asyncpg==0.30.0
httpx==0.27.2
beautifulsoup4==4.12.3
google-generativeai==0.8.3
scrapy==2.11.2
pytest==8.3.3
